Ref: AppDelegate 에서 API 호출 함수를 NetworkManager 로 분리
